.github/scripts/requirements.txt
tabulate
google-api-python-client>=2.115.0
google-auth>=2.28.1
google-auth-oauthlib>=1.2.0
requests>=2.28.0
